add explicit cast to fix compile error
"error: non-constant-expression cannot be narrowed from type 'int64_t' (aka 'long long') to 'uint32_t' (aka 'unsigned int') in initializer list [-Wc++11-narrowing]"

+++ examples/train-text-from-scratch/train-text-from-scratch.cpp
@@ -1788,7 +1788,10 @@ void write_tensor(struct llama_file * file, struct ggml_tensor * tensor) {
     const char * name = ggml_get_name(tensor);
     uint32_t name_len = strlen(name);
     uint32_t nd = tensor->n_dims;
-    uint32_t ne[4] = { tensor->ne[0], tensor->ne[1], tensor->ne[2], tensor->ne[3] };
+    uint32_t ne[4] = { (uint32_t)tensor->ne[0], 
+                       (uint32_t)tensor->ne[1], 
+                       (uint32_t)tensor->ne[2], 
+                       (uint32_t)tensor->ne[3] };
     file->write_u32(nd);
     file->write_u32(name_len);
     file->write_u32(tensor->type);
